Refactor inbox structure: move components to a new `components` directory, consolidate type imports, and update references accordingly
