+++ src/inbox/components/index.ts
@@ -1,0 +1,3 @@
+export * from './IterableInbox';
+export * from './IterableInboxEmptyState';
+export * from './IterableInboxMessageCell';
