+++ src/inbox/types/InboxImpressionRowInfo.ts
